requirements.txt
streamlit>=1.30
sqlalchemy>=2.0
pandas>=2.0
matplotlib>=3.7
bcrypt>=4.0
python-dateutil>=2.8
